requirements.txt
streamlit
python-dotenv
openai>=1.54.0